Move _convolve_data to utils to prevent circular import
